requirements.txt
django 
djangorestframework 
psycopg2-binary 
python-dotenv
openai